renaming the pleasework program
